Fixed URL bugs & menu bug
URL requests were only working for the first from the array of folders,
and serving the 404 page.txt file for everything else. They’re now
working properly, with the 404 page.txt only being served when it’s
supposed to be.

The menu has been fixed – whenever something didn’t meet the criteria,
it spat out the last menu item which did match it. using unset($var)
was the solution.